working on bluetooth

+++ scripts/z_modinst.sh
@@ -10,6 +10,11 @@ USBDISK=NXT_AND_X86
 echo "  MOUNT   /media/zefie/$USBDISK/system.img"
 mount -o loop -t ext4 /media/zefie/$USBDISK/system.img /home/zefie/tmp
 
+if [ "$(df -h | grep /home/zefie/tmp | wc -l)" -ne "1" ]; then
+	echo "  ERROR   Failed to mount system.img"
+	exit 1;
+fi
+
 echo "  REMOVE  Existing modules for: $KVERS"
 rm -r /home/zefie/tmp/lib/modules/$KVERS
 
@@ -17,6 +22,21 @@ make INSTALL_MOD_PATH=/home/zefie/tmp modules_install
 
 echo "  REMOVE  sound/soc/intel/common/snd-soc-sst-acpi.ko"
 rm /home/zefie/tmp/lib/modules/$KVERS/kernel/sound/soc/intel/common/snd-soc-sst-acpi.ko
+
+echo "  BUILD   external/rtl8723bs-bluetooth/hci_uart.ko"
+make SUBDIRS=external/rtl8723bs-bluetooth clean 2>&1 > /dev/null
+make SUBDIRS=external/rtl8723bs-bluetooth 2>&1 > /dev/null
+
+echo "  CHECK   external/rtl8723bs-bluetooth/hci_uart.ko"
+if [ ! -f "external/rtl8723bs-bluetooth/hci_uart.ko" ]; then
+	echo "  ERROR   Building rtl8723bs-bluetooth hci_uart failed"
+else
+	echo "  INSTALL drivers/bluetooth/hci_uart.ko"
+	cp external/rtl8723bs-bluetooth/hci_uart.ko /home/zefie/tmp/lib/modules/$KVERS/kernel/drivers/bluetooth/hci_uart.ko
+fi
+
+echo "  CLEAN   external/rtl8723bs-bluetooth/hci_uart.ko"
+make SUBDIRS=external/rtl8723bs-bluetooth clean 2>&1 > /dev/null
 
 echo "  DEPMOD  $KVERS"
 depmod $KVERS -b /home/zefie/tmp -a
